server/requirements.txt
Django==2.2.8
psycopg2-binary==2.8.2
dj-database-url==0.5.0
djangorestframework==3.10.3
requests
django-redis
gunicorn
django-filter==2.2.0
django-constance==2.4.0